+++ src/main/php/net/daringfireball/markdown/Markdown.class.php
@@ -143,7 +143,7 @@ class Markdown {
       static $def= ['(' => '()', '"' => '"', "'" => "'"];
       $title= trim($matches[3]);
       if ('' !== $title && 0 === strcspn($title, '(\'"')) {
-        $title= trim($title, $def[$title{0}]);
+        $title= trim($title, $def[$title[0]]);
       } else {
         $title= null;
       }
@@ -184,7 +184,7 @@ class Markdown {
     });
     $this->addHandler('/^\> /', function($lines, $matches, $result, $ctx) {
       $lines->resetLine($matches[0]);
-      $result->append($ctx->enter(new BlockquoteContext())->parse($lines));
+      $result->append($ctx->enter(new BlockquoteContext($ctx->handlers))->parse($lines));
       return true;
     });
     $this->addHandler('/^(    |\t)/', function($lines, $matches, $result, $ctx) {